Refactor balance type to BigDecimal and clean up code
The balance attribute in several areas of the code was updated to use BigDecimal instead of double for precision. In addition, some unnecessary comments and spaces were removed from Home.tsx for cleaner code. The changes apply to both frontend and backend portions of the code.

+++ backend/src/main/java/com/example/backend/entity/Account.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import jakarta.validation.constraints.Pattern;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -38,7 +39,7 @@ public class Account {
     @Column(length = 65,nullable=false)
     private String password;
 
-    private double balance;
+    private BigDecimal balance;
 
     @OneToMany(mappedBy = "account", cascade = CascadeType.ALL, orphanRemoval = true)
     @JsonManagedReference
